Causal Influence for 2 Coupled Lorenz Systems
Compared to the previous MLE_6D.py, the order of coefficients had to be change so that it aligns with the order in the Sympy-formulation. 
As of now, it's only up to the 2nd order polynomials. 3rd order may be a next step.
Works somewhat alright: the heatmap for the "Optimal Threshold Parameters" in block [18] is perfect. The heatmap for the naive estimation without a threshold needs some clever cutoff after calculating the causal influences, I think.