Deliberately wrong model for D2
Now I designed data with a D2 depending on x. However, a fully parametrised D2 = a + bx1 + cx2 + ... results in a mess for D1; probably because of the huge degree of freedom associated with such a generalised noise term. Hence, my idea is:
1) Estimate D1_est and a constant D2_0 (even although the latter is deliberately oversimplified!)
2) Use D1_est as *fixed* terms for the D1-part of the likelihood and NOW allow a more flexible model for D2_est (maybe with D2_0 as a benchmark in terms of BIC): This may reduce the degrees of freedom enough so that the likelihood estimation actually works

Step (1) is currently implemented in this code, but results are still pending.