MCMC for noise parameters
Estimating the noise via Maximum Likelihood for "known" drift parameters (estimated previously as BestCoeff[0][1:]) does not work: highly sensitive on initial guess and does not even come close to a good solution. That's probably because of the huge amount of degrees of freedom with so many (13) noise parameters. Instead, try to MCMC-sample the likelihood for the "known" drift terms. Hopefully, the superfluous terms have a broader distribution or peak around 0?